remove query param, update change log, move table into own comp
